Trigger animations based on state
Also switch Millis from int to long to allow for times longer than 32s

+++ global.h
@@ -1,7 +1,7 @@
 #ifndef __GLOBAL_H__
 #define __GLOBAL_H__
 
-typedef unsigned int Millis;
+typedef unsigned long Millis;
 typedef unsigned int Fps;
 
 #endif
